refactor/be/#247 sessionmanager core로 이동, 필요 없는 파일들 정리
